clean up of Icon and Motherdemon and shot the config spam up.

+++ src/main/java/mod/azure/doom/entity/projectiles/ArgentBoltEntity.java
@@ -213,7 +213,7 @@ public class ArgentBoltEntity extends AbstractArrowEntity {
 				((LivingEntity) entity1).setLastHurtMob(entity);
 			}
 		}
-		if (entity.hurt(damagesource, Config.SERVER.argent_bolt_damage)) {
+		if (entity.hurt(damagesource, Config.SERVER.argent_bolt_damage.floatValue())) {
 			if (entity instanceof LivingEntity) {
 				LivingEntity livingentity = (LivingEntity) entity;
 				if (!this.level.isClientSide && entity1 instanceof LivingEntity) {
