+++ src/main/java/mod/azure/doom/entity/projectiles/BFGEntity.java
@@ -226,7 +226,7 @@ public class BFGEntity extends AbstractArrowEntity implements IAnimatable {
 				if (d12 <= 1.0D) {
 					if (entity.isAlive()) {
 						entity.hurt(DamageSource.playerAttack((PlayerEntity) this.shooter),
-								Config.SERVER.bfgball_damage_aoe);
+								Config.SERVER.bfgball_damage_aoe.floatValue());
 						this.setTargetedEntity(entity.getId());
 					}
 				}
@@ -234,7 +234,7 @@ public class BFGEntity extends AbstractArrowEntity implements IAnimatable {
 			if (entity instanceof EnderDragonEntity) {
 				if (entity.isAlive()) {
 					((EnderDragonEntity) entity).head.hurt(DamageSource.playerAttack((PlayerEntity) this.shooter),
-							Config.SERVER.bfgball_damage_aoe);
+							Config.SERVER.bfgball_damage_aoe.floatValue());
 					this.setTargetedEntity(entity.getId());
 				}
 			}
@@ -327,7 +327,7 @@ public class BFGEntity extends AbstractArrowEntity implements IAnimatable {
 							|| entity instanceof HoglinEntity)) {
 				double d12 = (double) (MathHelper.sqrt(entity.distanceToSqr(vector3d)) / f2);
 				if (d12 <= 1.0D) {
-					entity.hurt(DamageSource.playerAttack((PlayerEntity) this.shooter), Config.SERVER.bfgball_damage);
+					entity.hurt(DamageSource.playerAttack((PlayerEntity) this.shooter), Config.SERVER.bfgball_damage.floatValue());
 					this.setTargetedEntity(entity.getId());
 					if (!this.level.isClientSide) {
 						List<LivingEntity> list1 = this.level.getEntitiesOfClass(LivingEntity.class,
@@ -352,7 +352,7 @@ public class BFGEntity extends AbstractArrowEntity implements IAnimatable {
 			if (entity instanceof EnderDragonEntity) {
 				if (entity.isAlive()) {
 					((EnderDragonEntity) entity).head.hurt(DamageSource.playerAttack((PlayerEntity) this.shooter),
-							Config.SERVER.bfgball_damage_dragon);
+							Config.SERVER.bfgball_damage_dragon.floatValue());
 				}
 			}
 		}
