+++ src/main/java/mod/azure/doom/entity/projectiles/BulletEntity.java
@@ -234,7 +234,7 @@ public class BulletEntity extends AbstractArrowEntity implements IAnimatable {
 				((LivingEntity) entity1).setLastHurtMob(entity);
 			}
 		}
-		if (entity.hurt(damagesource, Config.SERVER.bullet_damage)) {
+		if (entity.hurt(damagesource, Config.SERVER.bullet_damage.floatValue())) {
 			if (entity instanceof LivingEntity) {
 				LivingEntity livingentity = (LivingEntity) entity;
 				if (!this.level.isClientSide && entity1 instanceof LivingEntity) {
